% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{as.mpsth}
\alias{as.mpsth}
\title{Convert a list of STAR psth objects into an mpsth object}
\usage{
as.mpsth(x, ...)
}
\arguments{
\item{x}{list of psth objects}

\item{...}{ignored}
}
\value{
list of class mpsth
}
\description{
Convert a list of STAR psth objects into an mpsth object
}
\author{
jefferis
}
\seealso{
\code{\link{psth}}

Other mpsth: \code{\link{[.mpsth}},
  \code{\link{as.ts.mpsth}}
}

