% Generated by roxygen2 (4.0.1): do not edit by hand
\name{divide}
\alias{divide}
\alias{divide.spiketimes}
\title{Divide spiketimes object with multiple repeats into list with one entry per
repeat}
\usage{
divide(x, ...)

\method{divide}{spiketimes}(x, blocksize, ...)
}
\arguments{
\item{...}{Additional arguments (currently ignored)}

\item{x}{The spiketimes object}

\item{blocksize}{number of waves per block - deduced from odour names if
missing}
}
\value{
new spiketimes object (list) with one entry for each block of odours
}
\description{
divide.spiketimes is designed for the situation where a single pxp file
contains multiple repeats for the same stimulus set.
}
\details{
Only works for spiketimes from single pxp file. FIXME Teach this to
  cope with repeat blocks with different odours in different order
}
\author{
jefferis
}
\seealso{
Other spiketimes: \code{\link{+.spiketimes}};
  \code{\link{[.spiketimes}}; \code{\link{is.spiketimes}};
  \code{\link{merge.spiketimes}}; \code{\link{spiketimes}};
  \code{\link{split.spiketimes}};
  \code{\link{subset.spiketimes}}
}

