% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finder_colour}
\alias{finder_colour}
\title{Query the colour label of a file in the MacOS X Finder}
\usage{
finder_colour(filename)
}
\arguments{
\item{filename}{Path to a file}
}
\value{
Character string (None when no colour is set) or NA on error
}
\description{
Note the colours are mapped according to my 10.6.8 system but see:
http://stackoverflow.com/questions/2435580/tagging-files-with-colors-in-os-x-finder-from-shell-scripts
}
\seealso{
Other finder_colour: \code{\link{finder_colour_fast}},
  \code{\link{set_finder_colour}}
}
\author{
jefferis
}
