% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiketimes.R
\name{IgorWavesForSpikes}
\alias{IgorWavesForSpikes}
\title{Find the Igor waves matching a set of spikes (even if shuffled)}
\usage{
IgorWavesForSpikes(x)
}
\arguments{
\item{x}{a spiketimes object}
}
\description{
Find the Igor waves matching a set of spikes (even if shuffled)
}

