% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spike_analysis.R
\name{lifetime_sparseness}
\alias{lifetime_sparseness}
\title{Calculate lifetime sparseness according to Olsen et al. 2008}
\usage{
lifetime_sparseness(x, minodours = 0.75)
}
\arguments{
\item{x}{matrix of odour data with one column per odour, 1 row per cell}

\item{minodours}{Minimum number of odours required to compute valid score
(see details)}
}
\value{
named numeric vector, one entry for each row (neuron) in x
}
\description{
minodours is interpreted as fraction when <=1. If >1 it is assumed then there
must be >= that many odour responses. Any cells that have been tested for
fewer than than minodours will have a sparseness value of NA. Definition:
}
\details{
S = 1/(1-1/N)*( 1 - (sum(r/N)^2)/(sum(r^2/N)) )

where N is the number of odours and r is the analogue response intensity of
the neuron to odour j minus baseline firing rate. Analogue response intensity
was the mean spike rate (averaged across five sweeps) during the entire
500-ms odour stimulus period. Any values of rj < 0 were set to zero before
computing lifetime sparseness (this was the only analysis in this study in
which negative responses were zeroed). Responses to paraffin oil (solvent
control) were not considered in the sparseness analysis. We computed the
lifetime sparseness for each individual cell in our study for which we tested
at least 12 of the odours in our set. We used the Mann--Whitney U test to
assess the significance of sparseness differences.
}
\examples{
spikes=CollectSpikesFromSweeps(
  system.file('igor','spikes','nm20110914c4',package='gphys'),
  subdir='BLOCKI',sweeps=0:4)
od=OdourResponseFromSpikes(spikes,response=c(2200,2700),baseline=c(0,2000))
S=lifetime_sparseness(od)
}
\references{
Olsen S.R. and Wilson R.I. (2008). Lateral presynaptic inhibition
  mediates gain control in an olfactory circuit. Nature 452 (7190), 956--960.
  doi:10.1038/nature06864

B. Willmore, D. J. Tolhurst, Network Comput. Neural Syst. 12, 255
  (2001)
}

