% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiketimes.R
\name{MakeAverageWaves}
\alias{MakeAverageWaves}
\title{Make average (analogue) waves that match a spiketimes object}
\usage{
MakeAverageWaves(x, wavestem = "RecordA")
}
\arguments{
\item{x}{A spiketimes object}

\item{wavestem}{The stem of the wave names in Igor (defaults to 
\code{"RecordA"})}
}
\value{
A \code{mts} object with as many rows as there are Waves in the 
  stimulus protocol (i.e. normally odours).
}
\description{
You can use the \code{wavestem} argument to choose the analogue input channel
to use for calculation.
}
\seealso{
\code{\link{CollectSpikesFromSweeps}}
}
