% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split.spiketimes}
\alias{split.spiketimes}
\title{Split a spiketimes object with multiple repeats into list with one entry per repeat}
\usage{
\method{split}{spiketimes}(st, blocksize)
}
\arguments{
\item{st}{The spiketimes object}

\item{blocksize}{number of waves per block - deduced from odour names if missing}
}
\value{
new spiketimes object (list) with one entry for each block of odours
}
\description{
Only works for spiketimes from single pxp file
FIXME Teach this to cope with repeat blocks with different odours in different order
}
\author{
jefferis
}
\seealso{
Other spiketimes: \code{\link{+.spiketimes}};
  \code{\link{[.spiketimes}}; \code{\link{is.spiketimes}};
  \code{\link{merge.spiketimes}}; \code{\link{spiketimes}};
  \code{\link{subset.spiketimes}}
}

