% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiketimes.R
\name{subset.spiketimes}
\alias{subset.spiketimes}
\alias{subset}
\title{Subset spiketimes object to contain only sweeps for an odour subset}
\usage{
\method{subset}{spiketimes}(x, odours = NULL, channels = NULL, ...)
}
\arguments{
\item{x}{The original spiketimes object}

\item{odours}{A character vector of odours}

\item{channels}{Integer vector of channels}

\item{...}{further arguments to be passed to or from other methods (currently
ignored)}
}
\value{
spiketimes object restricted to specified odours or channels
}
\description{
NB the sweeps will be in the specified odour/channel order. If an unnamed
second parameter is specified it will be interpreted as vector of odours if
it is a character or channel ids if it is numeric.
}
\details{
If an odour name is duplicated in the oddconfig (e.g. ctr or oil, often) then
a warning will be given and the first matching sweep(s) will be used.
}
\seealso{
Other spiketimes: \code{\link{+.spiketimes}},
  \code{\link{[.spiketimes}}, \code{\link{divide}},
  \code{\link{is.spiketimes}},
  \code{\link{merge.spiketimes}}, \code{\link{spiketimes}},
  \code{\link{split.spiketimes}}
}
\author{
jefferis
}
\concept{spiketimes}
