\name{tsp.psth}
\alias{tsp.psth}
\title{Find start, end and frequency of a STAR psth object
Checks that it is regularly sampled}
\usage{
  tsp.psth(x)
}
\arguments{
  \item{x}{STAR psth object}
}
\value{
  numeric vector with start, end and frequency
}
\description{
  Find start, end and frequency of a STAR psth object
  Checks that it is regularly sampled
}
\seealso{
  tsp

  Other psth: \code{\link{as.ts.psth}}
}

