\name{as.mpsth}
\alias{as.mpsth}
\title{Convert a list of STAR psth objects into an mpsth object}
\usage{
as.mpsth(x, ...)
}
\arguments{
  \item{x}{list of psth objects}

  \item{...}{ignored}
}
\value{
list of class mpsth
}
\description{
Convert a list of STAR psth objects into an mpsth object
}
\author{
jefferis
}
\seealso{
\code{\link{psth}}

Other mpsth: \code{\link{[.mpsth}};
\code{\link{as.ts.mpsth}}
}

