% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiketimes.R
\name{divide}
\alias{divide}
\alias{divide.spiketimes}
\title{Divide spiketimes object with multiple repeats into list with one entry per 
repeat}
\usage{
divide(x, ...)

\method{divide}{spiketimes}(x, blocksize, ...)
}
\arguments{
\item{x}{The spiketimes object}

\item{...}{Additional arguments (currently ignored)}

\item{blocksize}{number of waves per block - deduced from odour names if 
missing}
}
\value{
new spiketimes object (list) with one entry for each block of odours
}
\description{
divide.spiketimes is designed for the situation where a single pxp file 
contains multiple repeats for the same stimulus set.
}
\details{
Only works for spiketimes from single pxp file.
  
  The function can cope with cases where odour order has been shuffled within
  repeated blocks each containing all test odours (but not when the odours
  could appear at any point in the test sequence). 
  
  For shuffled data the Wave column will be a number from 0 to blocksize-1
  that indicates which of the presented channels the spikes belong to. Note
  that this is sorted by absolute channel number \bold{not} by the order of
  presentation of odours in the first block.
  
  Likwise the \code{oddconf} of the returned list of data.frames will have
  the odours sorted in absolute channel order.
}
\seealso{
Other spiketimes: \code{\link{+.spiketimes}},
  \code{\link{[.spiketimes}}, \code{\link{is.spiketimes}},
  \code{\link{merge.spiketimes}}, \code{\link{spiketimes}},
  \code{\link{split.spiketimes}},
  \code{\link{subset.spiketimes}}
}
\author{
jefferis
}
\concept{spiketimes}
