\name{is.spiketimes}
\alias{is.spiketimes}
\title{Test if object is of class spiketimes}
\usage{
is.spiketimes(x)
}
\arguments{
  \item{x}{Object to test}
}
\value{
Logical indicating whether object is of class spiketimes
}
\description{
Test if object is of class spiketimes
}
\seealso{
Other spiketimes: \code{\link{+.spiketimes}};
\code{\link{[.spiketimes}}; \code{\link{merge.spiketimes}};
\code{\link{spiketimes}}; \code{\link{split.spiketimes}};
\code{\link{subset.spiketimes}}
}

