% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_analysis.R
\name{lifetime_sparseness}
\alias{lifetime_sparseness}
\title{Calculate lifetime sparseness according to Olsen et al. 2008}
\usage{
lifetime_sparseness(x, minodours = 0.75)
}
\arguments{
\item{x}{matrix of odour data with one column per odour, 1 row per cell}

\item{minodours}{Minimum number of odours required to compute valid score 
(see details)}
}
\value{
named numeric vector, one entry for each row (neuron) in x
}
\description{
Calculate lifetime sparseness according to Olsen et al. 2008
}
\details{
minodours is interpreted as fraction when <=1. If >1 it is assumed 
  then there must be >= that many odour responses. Any cells that have been 
  tested for fewer than than minodours will have a sparseness value of NA. 
  Definition:
  
  S = 1/(1-1/N)*( 1 - (sum(r/N)^2)/(sum(r^2/N)) )
  
  where N is the number of odours and r is the analogue response intensity of
  the neuron to odour j minus baseline firing rate. Analogue response 
  intensity was the mean spike rate (averaged across five sweeps) during the 
  entire 500-ms odour stimulus period. Any values of rj < 0 were set to zero 
  before computing lifetime sparseness (this was the only analysis in this 
  study in which negative responses were zeroed). Responses to paraffin oil 
  (solvent control) were not considered in the sparseness analysis. We 
  computed the lifetime sparseness for each individual cell in our study for 
  which we tested at least 12 of the odours in our set. We used the 
  Mann--Whitney U test to assess the significance of sparseness differences.
}
\section{Missing Values (NAs)}{
 When there are missing values in \code{x}, 
  they are omitted silently. NAs do not count towards the value of N, the 
  number of responses /cell in the equation above.
}

\examples{
spikes=CollectSpikesFromSweeps(
  system.file('igor','spikes','nm20120514c2',package='gphys'),
  subdir='BLOCKA')
od=OdourResponseFromSpikes(spikes,response=c(2200,2700),baseline=c(0,2000))
# NB we take mean response per odour since we must end up with one number
# per odour per cell
S=lifetime_sparseness(colMeans(od))
data.frame(odour=colnames(od), S=S)

# check what happens with NAs
od2=od
od2[5,'cVA']=NA
lifetime_sparseness(colMeans(od2))

# compare with a sparse cell
spikes2=CollectSpikesFromSweeps(
  system.file('igor','spikes','nm20110914c4',package='gphys'),
  subdir='BLOCKI', sweeps=0:4)
od3=OdourResponseFromSpikes(spikes2,response=c(2200,2700),baseline=c(0,2000))
lifetime_sparseness(colMeans(od3))
}
\references{
Olsen S.R. and Wilson R.I. (2008). Lateral presynaptic inhibition
  mediates gain control in an olfactory circuit. Nature 452 (7190), 956--960.
  \href{http://doi.org/10.1038/nature06864}{doi:10.1038/nature06864}
  
  Willmore B. and  Tolhurst D.J.(2001). Characterizing the sparseness of 
  neural codes. Network Comput. Neural Syst. 12, 255 (2001) 
  \href{http://doi.org/10.1080/net.12.3.255.270}{doi:10.1080/net.12.3.255.270}
}
