% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiketimes.R
\name{spiketimes}
\alias{spiketimes}
\title{Make spiketimes object (spike times + stimulus info) from list of dataframes}
\usage{
spiketimes(x, xlim, stimRange)
}
\arguments{
\item{x}{object to convert to spiketimes list (normally already a list(}

\item{xlim}{recording time window for each sweep}

\item{stimRange}{time window over which stimulus was delivered}
}
\value{
spiketimes object
}
\description{
spiketimes objects consist of a list of dataframes reporting the time of
spikes. the data frames have two core columns Time and Wave. Time is the time
in ms at which each spike occurred within the current sweep. Wave is the
0-indexed number of the sweep within the pxp file (Igor convention). Every
Wave must have an entry, so if there are no spikes in e.g. Wave 2, an entry
of (Time=NA,Wave=2) will be required.
}
\details{
Separate dataframes can be combined into a list where each dataframe is one 
block of waves that is repeated within a single pxp file or compatible blocks
from multiple pxp files. Use \code{\link{split.spiketimes}} to split a
spiketimes object loaded from a pxp file with repeated blocks and 
\code{\link{+.spiketimes}} to combine compatible blocks.

\code{\link{CollectSpikesFromSweeps}} will generate a spiketimes object for 
you, so this function will only needed if you are constructing 
\code{\link{spiketimes}} objects from scratch.
}
\seealso{
Other spiketimes: \code{\link{+.spiketimes}},
  \code{\link{[.spiketimes}}, \code{\link{divide}},
  \code{\link{is.spiketimes}},
  \code{\link{merge.spiketimes}},
  \code{\link{split.spiketimes}},
  \code{\link{subset.spiketimes}}
}
\concept{spiketimes}
