% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spiketimes.R
\name{split.spiketimes}
\alias{split.spiketimes}
\title{Split a spiketimes object with multiple repeats into list with one entry per
repeat}
\usage{
\method{split}{spiketimes}(x, f, drop, ...)
}
\arguments{
\item{x}{The spiketimes object}

\item{f}{number of waves per block - deduced from odour names if missing}

\item{drop}{IGNORED}

\item{...}{IGNORED}
}
\value{
new spiketimes object (list) with one entry for each block of odours
}
\description{
\emph{Deprecated} as of gphys 0.9 and will be removed from gphys 1.0 (CRAN
release).
}
\details{
Only works for spiketimes from single pxp file. FIXME Teach this to cope with
repeat blocks with different odours in different order
}
\author{
jefferis
}
\seealso{
Other spiketimes: \code{\link{+.spiketimes}};
  \code{\link{[.spiketimes}}; \code{\link{divide}},
  \code{\link{divide.spiketimes}};
  \code{\link{is.spiketimes}};
  \code{\link{merge.spiketimes}}; \code{\link{spiketimes}};
  \code{\link{subset.spiketimes}}
}

